Add Warp module
An initial, unfinished implementation of a simple audio signal warper. It's kind of analogous to a bus/send in other DAWs.

Any input signals will be "transmitted" into a shared buffer for a specified identifier, and any other modules with that same identifier will "receive" that signal and output it.