Only keep last output per ident, track all patchers for Warps
Simplifies the code a lot and makes this less of a state machine mess - although, it still is. Not every output needs to be tracked, only the last output of each ident.

There was also an issue related to an input with multiple inputting patchers, so mPatcher was changed to a vector.